User Cookies
I made it so the user is saved in the cookies and that the cookie value will be used to make calls to the backend instead of having path parameters on the frontend and using those parameters to make API calls. I also made sure that for user-specific pages, that that user must be logged in.