Started The Class Homepage
I started making all the JSX components that will be used in the classroom's pages. I have only started work on the frontend of this. I still need to make the backend. The backend will be done AFTER making the math section so that math problems may be assigned and what not.