Made more backend code for the assignments
I made the backend for the assignments so that they could be created. When an assignment is created, it creates a assignmentStudentDetails object to which is stored in its own DB and a reference is stored in the Assignment object. This is because this extra data is only needed a small number of times when you need Assignment Data, and so, to reduce load time when fetching assignments, this extra data is only fetched when needed not every time an Assignment is fetched.

I also made it so problem sets can be generated from the math_router in the backend.