Made a number Navbar
Can be used to select a number index from a range of indices. Also made some other changes to the assignment pages for educators and students to load the questions/answers with the createResources function. Its useful when I have a createEffect that uses a resource that needs to be fetched.